First Dotfile Upload from WSL2 Ubuntu 20.04
This is more of an experiment on my behalf, trying to figure out
how exactly to work the dotfile sharing side of yadm and also
how to eventually start syncing and working on the CLI workflow.

Especially as I consider getting back into TAFE Cert IV, it's
imperative that I continue to sharpen my CLI skills as it's
a major skill in the networks admin industry in which I hope
to gain employment after my mental health issues are sorted.